Move `FloatingBaseMenu` to its own file
